MpegTools : Mises à jour de compatibilité C++11 dans “TsChecker” et mise à jour du numéro de révision.
git-svn-id: http://svn.pouchintv.fr/pouchintv/trunk/MpegTools@378 2c5fd688-289b-411d-a95c-3d4e4b52ee9e

+++ TsChecker/TsCheckerDlg.cpp
@@ -209,7 +209,7 @@ void CTsCheckerDialog::ShowFileInfo(UINT nId, LPCTSTR pszFileName)
 	WIN32_FILE_ATTRIBUTE_DATA sData;
 
 	if (GetFileAttributesEx(pszFileName, GetFileExInfoStandard, &sData)) {
-		LARGE_INTEGER	sSize = {sData.nFileSizeLow, sData.nFileSizeHigh};
+		LARGE_INTEGER	sSize = {sData.nFileSizeLow, static_cast<LONG>(sData.nFileSizeHigh)};
 		TCHAR			szWrk[64];
 
 		_stprintf_s(szWrk, _TL("Size = %s Kb","Taille = %s Ko"), thousandsSepFmt((sSize.QuadPart+0x200) / 0x400).c_str());
